+++ packages/system/src/components/Formatters/index.ts
@@ -1,2 +1,2 @@
-export * from './Date';
+export * from './DateFormat';
 export * from './NumericFormat';
